Remove deprecated agent backup file from single_file_agent directory
